Add e2e test for android deployment and launch tasks.
ASwB's run/debug area is missing tests that cover the overall
flow from creation of a run configuration to the generation of
APK deployment and launch tasks. There exist tests that verify
the individual steps through out this flow but breakages in the
overall integration can't be caught by these existing tests.

This CL adds an end-to-end test with coverage from the creation of
a run configuration to the generation of the APK deployment
and launch tasks.  Note this CL only adds one test and it
only covers running an android_binary target with mobile-install
disabled.  This CL serves as the "first of its kind" and more
tests are required for better coverage.

PiperOrigin-RevId: 328298176